refactor(lotto-factory):로또 생성에 대한 책임을 가진 클래스 정의
