+++ src/main/java/lotto/domain/LottoFactory.java
@@ -1,0 +1,18 @@
+package lotto.domain;
+
+import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+
+public class LottoFactory {
+    
+    private LottoFactory() {}
+    
+    public static Lottos makeLottos(PaymentAmount paymentAmount) {
+        List<Lotto> lottos = Stream.generate(() -> Lotto.fromGenerator(new RandomNumberGenerator()))
+                                   .limit(paymentAmount.getPurchaseCount())
+                                   .collect(Collectors.toList());
+        
+        return new Lottos(lottos);
+    }
+}
